Updated with PR Comments
