+++ Scripts/Runtime/Entities/TaskSystem/TaskFlow/TaskFlowSystem.cs
@@ -14,6 +14,8 @@ namespace Anvil.Unity.DOTS.Entities.Tasks
             get;
         }
 
+        private bool m_HasInitialized;
+
         public TaskFlowSystem()
         {
             TaskFlowGraph = new TaskFlowGraph();
@@ -22,6 +24,14 @@ namespace Anvil.Unity.DOTS.Entities.Tasks
         protected override void OnStartRunning()
         {
             base.OnStartRunning();
+            
+            //If for some reason this System gets re-enabled we don't want to initialize the graph anymore.
+            if (m_HasInitialized)
+            {
+                Enabled = false;
+                return;
+            }
+            m_HasInitialized = true;
             
             TaskFlowGraph.ConfigureTaskSystemJobs();
             //TODO: #68 - Probably a better way to do this via a factory type. https://github.com/decline-cookies/anvil-unity-dots/pull/59#discussion_r977823711
